Adds new `ExternalLink` component and updates changelog
- Created new `ExternalLink` component that is to be used wherever a link to an external web page is needed. The new component works differently depending on if the Web or Desktop version of the application is being used.
- Updated changelog to include note about new `PostFocusModal` "reply container scroll position save/restore" feature.
