+++ src/components/Settings/AboutAppModal.vue
@@ -38,6 +38,13 @@
             <div class="flex flex-col overflow-y-auto pt-2">
                 <div class="px-4 text-xl font-semibold">Changelog:</div>
                 <div class="flex flex-col gap-1 h-full px-4 divide-outline divide-y">
+                    <div class="font-semibold">March 13th 2026</div>
+                    <ul class="list-disc list-inside h-full py-1 text-sm">
+                        <li>Updated app so that the scroll position is saved and restored when navigating a Post Thread with the "Post Focus Modal" (<ExternalLink link-url="https://github.com/sondaismith/moongate/pull/286" link-text="#286"/>).
+                        </li>
+                    </ul>
+                </div>
+                <div class="flex flex-col gap-1 h-full px-4 divide-outline divide-y">
                     <div class="font-semibold">February 10th 2026</div>
                     <ul class="list-disc list-inside h-full py-1 text-sm">
                         <li>Updated app so that application state is synced between all open web-based application instances.
@@ -155,13 +162,7 @@
                 </div>
             </div>
             <div class="flex gap-1 px-4 py-2 text-xs">
-                <div>Any issues? Report</div>
-                <a v-if="!isTauri()" target="_blank"
-                :href="issuesURL"
-                class="text-blue-500 hover:text-blue-300 cursor-pointer">here!</a>
-                <a v-else
-                @click="(e) => showOptionsMenu(e,issuesURL)"
-                class="text-blue-500 hover:text-blue-300 cursor-pointer">here!</a>
+                <div>Any issues? Report <ExternalLink :link-url="issuesURL" link-text="here!"/></div>
             </div>
         </div>
     </div>
@@ -179,14 +180,7 @@ import { IOptionMenuItem, ItemType } from '../Utilities/OptionsMenu.vue';
 import { openUrl } from '@tauri-apps/plugin-opener';
 import AppLogo from '../SVG/AppLogo.vue';
 import { GetVersion, IVersionDetails } from '../../lib/api/VersionService';
-
-/**
- * Method used to open link in the system's default browser.
- * @param url The URL to open in the default browser.
- */
-async function OpenLink(url:string){
-    await openUrl(url);
-}
+import ExternalLink from '../Utilities/ExternalLink.vue';
 
 export default defineComponent({
     props:{
@@ -196,7 +190,8 @@ export default defineComponent({
         }
     },
     components:{
-        AppLogo
+        AppLogo,
+        ExternalLink,
     },
     data(){
         return{
@@ -220,18 +215,6 @@ export default defineComponent({
         },
         closeModal(){
             this.$router.push('/');
-        },
-        /**
-         * Shows Options Menu allowing user to perform different actions
-         * relating to the selected link.
-         */
-        showOptionsMenu(e:MouseEvent, linkURL:string){
-            e.preventDefault();
-            OptionsMenuState.currentMenuItems = [
-                {Icon:MingcuteWorld2Line,Label:'Open in Default Browser',Action:function(){OpenLink(linkURL)},Type:ItemType.Option},
-                {Icon:MingcuteCopyLine,Label:'Copy link to clipboard',Action:function(){CopyTextToClipboard(linkURL,'link')},Type:ItemType.Option},
-            ] as IOptionMenuItem[]
-            OptionsMenuState.showOptionMenu(e);
         },
     },
     created() {
